Saturation improvements and others
SaturationStateGivenDeltaRRND() updater now correctly indexes all
unknowns following a NewtonUpdater() contraction.

The new formulation has resulted in new stability limits near the
critical point.

UniqueEnough() was added such that values that differ by less than the
iteration tolerance will be treated as unique.

VectorChunk() splits up a given vector as speified by the chunk size
argument.

none() return true is all element along a particular dimension are
false.

every() return true is every element of an array is true.  It is the
same as all() with the input columnified.

notevery() is the opposite of every.